LSP: make --lsp-max-files-on-fast-path overflow-safe and describe it as a budget
Adversarial review of the previous commit:

- `fastPathFilesToTypecheck` compared against `2 * lspMaxFilesOnFastPath` in
  `uint32_t`; a value of 2^31 or more wrapped that to a tiny budget, so the
  dependents scan stopped early while the fast-path decision (made against the
  unwrapped value) still accepted the edit -- leaving dependents untypechecked.
  Do the multiplication in 64 bits. `options_test` checks the largest value
  round-trips.
- The help text and `lsp.md` read as if fitting the budget were sufficient for
  the fast path; new files, package files, syntax errors and changed
  definitions force the slow path regardless. Say so.
- `options.h` still described the field as "Not exposed on CLI".

Committed-By-Agent: claude

+++ main/lsp/LSPFileUpdates.cc
@@ -169,7 +169,8 @@ LSPFileUpdates::fastPathFilesToTypecheck(const core::GlobalState &gs, const LSPC
         result.extraFiles.emplace_back(ref);
         result.totalChanged += 1;
 
-        if (result.totalChanged > (2 * config.opts.lspMaxFilesOnFastPath)) {
+        // (64-bit so that a very large --lsp-max-files-on-fast-path cannot wrap this to a tiny budget.)
+        if (result.totalChanged > 2 * static_cast<uint64_t>(config.opts.lspMaxFilesOnFastPath)) {
             // Short circuit, as a performance optimization.
             // (gs.getFiles() is usually 3-4 orders of magnitude larger than lspMaxFilesOnFastPath)
             //
